Move Export and Import commands to menu bar
